+++ AlarmClock/Views And Controllers/MainScreenViewController.swift
@@ -224,7 +224,7 @@ class MainScreenViewController: UIViewController, UIPickerViewDelegate, UIPicker
     /// This is the basic background color for the whole kit and kaboodle. It gets darker as the brightness is reduced.
     var backgroundColor: UIColor = UIColor.gray
     /// This is the "heartbeat" of the clock. It's a 1-second repeating timer.
-    var timer: Timer!
+    var timer: RVS_BasicGCDTimer!
     /// This is used to cache the selected main font size. We sort of use it as a semaphore.
     var fontSizeCache: CGFloat = 0
     /// This contains information about music items.